Update Triggers.sql
Commented the UpdateArStatus_AfterInsertReviewLog trigger, as it is handled by the procedure InsertReviewLog, this procedure can get input of type by user status, so we can correctly update the new status for the article

Meanwhile, When insert an Edit log. the only next status is 'Wait', so a trigger can handle this truong hop.